# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
security:
  - cookieAuth: []
...
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get the logged-in user's own profile
      responses:
        "200":
          description: Own profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update the logged-in user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteMyAccount
      tags: [users]
      summary: Permanently delete the logged-in user's account
      responses:
        "204":
          description: Account deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{userId}:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get another user's public profile
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------- pets --
...
  /blocks:
    get:
      operationId: listBlockedUsers
      tags: [blocks]
      summary: List members the logged-in user has blocked
      responses:
        "200":
          description: Blocked members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: blockUser
      tags: [blocks]
      summary: Block a member, ending any match between them
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUserRequest"
      responses:
        "204":
          description: Member blocked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /blocks/{userId}:
    parameters:
      - $ref: "#/components/parameters/userId"
    delete:
      operationId: unblockUser
      tags: [blocks]
      summary: Lift a block
      responses:
        "204":
          description: Block lifted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # --------------------------------------------------------------- reports --
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pawmate_session
  parameters:
...
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Not signed in
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    LookingFor:
      type: string
      enum: [friendship, relationship, playdates, casual, open]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        age:
          type: integer
          nullable: true
        city:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        lookingFor:
          type: array
          items:
            $ref: "#/components/schemas/LookingFor"
        lifestyleTags:
          type: array
          items:
            type: string
        isOnline:
          type: boolean
        lastActiveAt:
          type: string
          format: date-time
          nullable: true
        twoFactorEnabled:
          type: boolean
        language:
          type: string
        onboardingCompletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lookingFor
        - lifestyleTags
        - isOnline
        - twoFactorEnabled
        - language
        - createdAt
        - updatedAt
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        age:
          type: integer
        city:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        lookingFor:
          type: array
          items:
            $ref: "#/components/schemas/LookingFor"
        lifestyleTags:
          type: array
          items:
            type: string

    # ------------------------------------------------------------- pets --
...
    BlockedUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        blockedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - blockedAt
    BlockUserRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
      required:
        - userId
...