# lib/api-zod/src/generated/types/blockedUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlockedUser {
  id: string;
  firstName: string;
  avatarUrl?: string | null;
  blockedAt: Date;
}